Fix a PHP 8 fatal when the recover form renders its captcha
osc_show_recaptcha('recover_password') read 'recover_time' out of the session and
subtracted it from time(). That key is only written by the POST action in
UserActions::recover_password(), and Session::_get() returns '' for a missing key,
so the first GET of the form computed `time() - ''` — a TypeError on PHP 8, and a
500 on the page. Casting to int is what ItemActions already does with
'last_submit_item' for the same reason.

Behaviour is otherwise unchanged: an unset key reads as 0, which falls outside the
20-minute window, so the branch records that no captcha was rendered and
recover_password() correctly skips validating one. A first visit to the form has
never shown a captcha; whether it should is a separate question.

Only bender reaches this — it is the one caller passing 'recover_password'. The
admin's own recover form calls osc_show_captcha() with no argument and was never
affected.

Fixes #5

+++ oc-includes/osclass/helpers/hUtils.php
@@ -116,8 +116,12 @@ function osc_show_widgets_by_description($description)
 
 
 /**
- * Print recaptcha html, if $section = "recover_password"
- * set 'recover_time' at session.
+ * Print recaptcha html.
+ *
+ * The 'recover_password' section only shows a captcha when a recovery attempt was made in
+ * the last 20 minutes; otherwise it records that none was rendered, so
+ * UserActions::recover_password() knows not to validate one. A first visit to the form
+ * therefore gets no captcha.
  *
  * @param string $section
  *
@@ -129,7 +133,9 @@ function osc_show_recaptcha($section = '')
         switch ($section) {
             case ('recover_password'):
                 Session::newInstance()->_set('recover_captcha_not_set', 0);
-                $time = Session::newInstance()->_get('recover_time');
+                // Cast: 'recover_time' is only set by the POST action, and Session::_get()
+                // returns '' for a missing key — subtracting that is a TypeError on PHP 8.
+                $time = (int)Session::newInstance()->_get('recover_time');
                 if ((time() - $time) <= 1200) {
                     echo _osc_recaptcha_get_html(osc_recaptcha_public_key(), substr(osc_language(), 0, 2)) . '<br />';
                 } else {
